requirements.txt
langchain
openai
python-dotenv
unstructured<=0.7.12
faiss-cpu
tiktoken
pyinstaller